+++ src/test/java/chess/dao/ChessDaoTest.java
@@ -4,7 +4,10 @@ import chess.domain.chessgame.ChessGame;
 import chess.domain.piece.Piece;
 import chess.domain.piece.Team;
 import chess.domain.position.Position;
-import org.junit.jupiter.api.*;
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
 
 import java.sql.*;
 import java.util.Map;
@@ -13,50 +16,48 @@ import java.util.stream.Collectors;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 
-class ChessDAOTest {
+class ChessDaoTest {
     private final String gameId = "new_game";
     private final ChessGame game = new ChessGame();
-    private final TestDAO chessDAO = new TestDAO();
+    private final TestDao chessDao = new TestDao();
 
     @BeforeEach
     public void setup() throws SQLException {
-        chessDAO.addGame(gameId, new ChessGame());
+        chessDao.addGame(gameId, new ChessGame());
     }
 
     @AfterEach
     void tearDown() throws SQLException {
-        chessDAO.deleteGameById(gameId);
+        chessDao.deleteGameById(gameId);
     }
 
     @Test
     @DisplayName("db 연결 기능")
     public void connection() {
-        Connection con = chessDAO.getConnection();
+        Connection con = chessDao.getConnection();
         assertNotNull(con);
     }
 
     @Test
     @DisplayName("게임 가져오는 기능")
-    @Order(3)
     void findGame() throws SQLException {
-        ChessGame gameFromData = chessDAO.findGameById(gameId);
-        assertThat(chessDAO.stringifiedBoard(game.board())).isEqualTo(chessDAO.stringifiedBoard(gameFromData.board()));
-        assertThat(chessDAO.stringifiedTurn(game.currentTurn())).isEqualTo(chessDAO.stringifiedTurn(gameFromData.currentTurn()));
+        ChessGame gameFromData = chessDao.findGameById(gameId);
+        assertThat(chessDao.stringifiedBoard(game.board())).isEqualTo(chessDao.stringifiedBoard(gameFromData.board()));
+        assertThat(chessDao.stringifiedTurn(game.currentTurn())).isEqualTo(chessDao.stringifiedTurn(gameFromData.currentTurn()));
     }
 
     @Test
     @DisplayName("게임 상태 업데이트 기능")
-    @Order(4)
     void updateGame() throws SQLException {
         game.move(new Position("b2"), new Position("b3"));
-        chessDAO.updateGame(gameId, game);
-        ChessGame gameFromData = chessDAO.findGameById(gameId);
-        assertThat(chessDAO.stringifiedBoard(game.board())).isEqualTo(chessDAO.stringifiedBoard(gameFromData.board()));
-        assertThat(chessDAO.stringifiedTurn(game.currentTurn())).isEqualTo(chessDAO.stringifiedTurn(gameFromData.currentTurn()));
+        chessDao.updateGame(gameId, game);
+        ChessGame gameFromData = chessDao.findGameById(gameId);
+        assertThat(chessDao.stringifiedBoard(game.board())).isEqualTo(chessDao.stringifiedBoard(gameFromData.board()));
+        assertThat(chessDao.stringifiedTurn(game.currentTurn())).isEqualTo(chessDao.stringifiedTurn(gameFromData.currentTurn()));
     }
 }
 
-final class TestDAO {
+final class TestDao {
     public Connection getConnection() {
         Connection con = null;
         String server = "localhost:13306"; // MySQL 서버 주소
@@ -100,13 +101,15 @@ final class TestDAO {
             throw new IllegalArgumentException("이미 존재하는 게임입니다.");
         }
         String query = "INSERT INTO games VALUES (?, ?, ?)";
-        Connection connection = getConnection();
-        PreparedStatement pstmt = connection.prepareStatement(query);
-        pstmt.setString(1, gameId);
-        pstmt.setString(2, stringifiedBoard(game.board()));
-        pstmt.setString(3, stringifiedTurn(game.currentTurn()));
-        pstmt.executeUpdate();
-        closeConnection(connection);
+        try (Connection connection = getConnection();
+             PreparedStatement pstmt = connection.prepareStatement(query)
+        ) {
+            pstmt.setString(1, gameId);
+            pstmt.setString(2, stringifiedBoard(game.board()));
+            pstmt.setString(3, stringifiedTurn(game.currentTurn()));
+            pstmt.executeUpdate();
+        } catch (SQLException ignored) {
+        }
     }
 
     String stringifiedBoard(Map<Position, Piece> board) {
@@ -122,46 +125,56 @@ final class TestDAO {
 
     public boolean isExistingGame(String gameId) throws SQLException {
         String query = "SELECT * FROM games WHERE game_id = ?";
-        Connection connection = getConnection();
-        PreparedStatement pstmt = connection.prepareStatement(query);
-        pstmt.setString(1, gameId);
-        ResultSet rs = pstmt.executeQuery();
-        boolean isExisting = rs.next();
-        closeConnection(connection);
-        return isExisting;
+        try (Connection connection = getConnection();
+             PreparedStatement pstmt = connection.prepareStatement(query)
+        ) {
+            pstmt.setString(1, gameId);
+            try (ResultSet rs = pstmt.executeQuery()) {
+                return rs.next();
+            }
+        } catch (SQLException throwables) {
+            return false;
+        }
     }
 
-    public ChessGame findGameById(String gameId) throws SQLException {
+    public ChessGame findGameById(String gameId) {
         String query = "SELECT * FROM games WHERE game_id = ?";
-        Connection connection = getConnection();
-        PreparedStatement pstmt = connection.prepareStatement(query);
-        pstmt.setString(1, gameId);
-        ResultSet rs = pstmt.executeQuery();
-        if (!rs.next()) {
-            throw new IllegalArgumentException("존재하지 않는 게임입니다.");
+        try (Connection connection = getConnection();
+             PreparedStatement pstmt = connection.prepareStatement(query)
+        ) {
+            pstmt.setString(1, gameId);
+            try (ResultSet rs = pstmt.executeQuery()) {
+                if (!rs.next()) {
+                    throw new IllegalArgumentException("존재하지 않는 게임입니다.");
+                }
+                return new ChessGame(rs.getString("pieces"), rs.getString("turn"));
+            }
+        } catch (SQLException throwables) {
+            return null;
         }
-        ChessGame game = new ChessGame(rs.getString("pieces"), rs.getString("turn"));
-        closeConnection(connection);
-        return game;
     }
 
     public void deleteGameById(String gameId) throws SQLException {
         String query = "DELETE FROM games WHERE game_id = ?";
-        Connection connection = getConnection();
-        PreparedStatement pstmt = connection.prepareStatement(query);
-        pstmt.setString(1, gameId);
-        pstmt.executeUpdate();
-        closeConnection(connection);
+        try (Connection connection = getConnection();
+             PreparedStatement pstmt = connection.prepareStatement(query)
+        ) {
+            pstmt.setString(1, gameId);
+            pstmt.executeUpdate();
+        } catch (SQLException ignored) {
+        }
     }
 
     public void updateGame(String gameId, ChessGame game) throws SQLException {
         String query = "UPDATE games SET pieces = ? , turn = ? WHERE game_id = ?";
-        Connection connection = getConnection();
-        PreparedStatement pstmt = connection.prepareStatement(query);
-        pstmt.setString(1, stringifiedBoard(game.board()));
-        pstmt.setString(2, stringifiedTurn(game.currentTurn()));
-        pstmt.setString(3, gameId);
-        pstmt.executeUpdate();
-        closeConnection(connection);
+        try (Connection connection = getConnection();
+             PreparedStatement pstmt = connection.prepareStatement(query)
+        ) {
+            pstmt.setString(1, stringifiedBoard(game.board()));
+            pstmt.setString(2, stringifiedTurn(game.currentTurn()));
+            pstmt.setString(3, gameId);
+            pstmt.executeUpdate();
+        } catch (SQLException ignored) {
+        }
     }
 }
